Use newer trash icon on Eclipse Photon or newer
Copy the SVG and PNGs from the eclipse.platform.images git repository
and rename them from "trash.*" to "trash_flat.*".

Add a version check for the first version of the "org.eclipse.ui"
bundle that uses the new icon to decide which one to use in EGit.
EGit UI has a hard (non-optional) dependency on org.eclipse.ui anyway,
so this version check is acceptable.

This ensures that EGit uses a trash icon visually identical to the
one used by the platform on which EGit runs, and thus the user is
presented with a consistent UI.

Bug: 525600
Change-Id: I98c869b0b84677bc67575edd7358a0a17954727c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>