Update patch set 4
Patch Set 4:

> Just a side comment without really looking at the code: I am
 > surprised that we have to do so much regarding smudge/clean filters
 > in EGit itself. I hoped that applying filters to content is JGits
 > business and I tell JGit only whether I want to checkout or checkin
 > so it's clear whether to smude or clean.

Correct metadata is available only during a treewalk. So I tried to avoid extra tree walks as much as possible (there are still a few), which means I have to pass through the metadata from the place where it was gotten to the place where the blob is read (or written). Or did I misunderstand something?

Patch-set: 4