Update patch set 1
Patch Set 1:

> Change itself looks good but I see no differences in the popup icon
 > on the repository. AFAICS the png is still hard-code via the
 > plugin.xml
 > 
 > <image
 > commandId="org.eclipse.egit.ui.team.GarbageCollect"
 > icon="icons/elcl16/trash.png">
 > </image>

Missed that one. Making this one dynamic depending on the org.eclipse.ui version will require a new property tester, and using it in the menu contribution.

Pfff... 

The others (that are changed by this implementation) are visible for instance in the general "Push..." dialog, table at the bottom where the push refspecs are listed, rightmost column.

Patch-set: 1