Update patch set 1
Patch Set 1:

> I must admit I have not understood the error message at all before
 > your posting. I used the RC version as well as the release in our
 > company project without any problem. Unfortunately I was not able
 > to maven build the egit project even before submitting this change
 > due to target platform trouble.

running the build using the latest version 3.5.2 locally works fine. 

According to
 $ mvn versions:display-plugin-updates
we need at least maven 3.3.1 required by eclipse-jarsigner-plugin 1.1.4.
We declare we need at least 3.0 which is too low, though we run the build using maven 3.3.9 on jenkins so this should work but it doesn't. Looks like a bug. 

I installed the latest maven 3.5.2 on jenkins, this seems to fix the issue.

Patch-set: 1