Replace FileInputStream and FileOutputStream with static Files methods
FileInputStream and FileOutputStream rely on finalize() method to ensure
resources are closed. This implies they are added to the finalizer queue
which causes additional work for the JVM GC process.

This is an open bug on the OpenJDK [1] and the recommended workaround is
to use the Files.newInputStream and Files.newOutputStream static methods
instead.

[1] https://bugs.openjdk.java.net/browse/JDK-8080225

Change-Id: I59e72d0a12d70396a5cf558029c13e2a1b0d5741
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>