Fix non-modal PushResultDialog
Commit f89d206 in 2015 was the first commit trying to make this dialog
non-modal. However, just clearing the SWT.APPLICATION_MODAL flag is
not enough: one also needs to set the dialog to not block on open,
i.e., not to run an event dispatching loop.

While I did not succeed to write an SWTbot unit test for bug 521176
and bug 530157, I'm pretty sure that this omission is what causes
the UI deadlock or livelock. We end up with two event dispatching
loops, one in this dialog, and the one of the modal dialog opened
by busyCursorWhile().

Calling setBlockOnOpen(false) for a non-modal PushResultDialog should
fix this.

Bug: 521176
Change-Id: I4626ced3fffe425807602fad27d4f9e7b71d189d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>