Fix focus handling in GitHistoryPage
GitHistoryPage.setFocus() unconditionally and always set the focus
to the commit graph table. Somehow that caused the focus tracking
for the various global actions (copy & selectAll) that are registered
and unregistered on focus changes to be thrown off when some action
in the CommitFileDiffViewer activated some other view or editor, and
focus then came back to the CommitMessageViewer.

So use the mechanism used already in the CommitEditor: remember which
control last had the focus, and if possible focus that one in
GitHistoryView.setFocus().

This fixes the handler conflicts we got that caused copy to no longer
work. I also verified manually that bug 322751 and bug 477510 do not
re-appear, and that the repository switch still works.

Bug: 503198
Change-Id: Id5d79f95836a50c43f6fbec477c5751a49cd9bdb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>