Include local branch name in proposals
If "Push branch 'foo'" is run the default suggestion for the
upstream name is the configured remote tracking branch (if any).
The local branch name is suggested initially only if no remote
tracking branch is set.

A frequent use case for this dialog is to push a local branch to
a *new* remote branch with the same name as the local branch. But
since typically local branches are created off 'master' and do
have 'origin/master' set as tracking branch, a user has in this
case to type in the new upstream branch name. It wouldn't appear
in the proposals list, which showed only existing upstream refs.

Simplify this use case by including a ref for a new upstream branch
in the list of proposals if there is no existing upstream branch
with the same name as the local branch.

Bug: 530685
Change-Id: I861d0ab73ed9476e664e5fe203da3a84829ce04f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>