Update patch set 19
Patch Set 19:

> Often LFS storage is used to version large binary files which would
 > first need support for custom diff drivers which we don't have yet
 > in JGit/EGit.

Well. The LFS changes over in EGit do add a kind of custom LFS diff driver that downloads the blob, if needed, and then does the normal raw-text diff on it.

 > Diff-ing a binary file in a text based diff tool
 > doesn't seem to be useful.

Yes, somewhere in all this there should be a check that the binary attribute isn't set for the item being handled.

Patch-set: 19