Update patch set 7
Patch Set 7:

The user needs some knob somewhere to control for which files smudging happens. It is the correct thing to do in the general case, but with LFS (and possibly other potentially long-running or otherwise expensive filters), the user may not want to have this for some files.

I don't see an obvious knob in standard git for this. Git config merge.renormalize doesn't fit the bill; it's a global setting applying to the whole repo and also is for a different use case (attributes differ between ours, their, and base). Git attribute -diff also isn't quite right, and might conflict with whatever the setting of the diff attribute is for command-line use. (Though perhaps we might want to disable comparisons in Eclipse if indeed -diff is active? -- But that'd be a different story.)

Maybe a different, new gitattribute? +rawdiff could mean that EGit shall not do the smudging. Th attribute could be passed along in the CheckoutMetadata. That would give a user the flexibility to suppress the smudging for individual files as needed. For instance, setting it on really large LFS files, but not, say, on LFS files that are text and smaller than some reasonable size.

Thoughts?

Patch-set: 7