Update patch set 7
Patch Set 7:

I would agree if we somehow force a diff on users (do we? do we automagically diff in some place?). As long as it is a user action to /request/ a diff for a file, I think we can offload at least some responsibility to the user. We should have kind of an emergency break if things get out of hand (e.g. timeout, RAM limit, etc.) so that Eclipse doesn't die due to large blobs. The rest can be "what would you expect if you request a diff on this file?" type of question to the user :)

But that doesn't mean I'm strictly against any more user friendliness if you think it is required - I'm just not yet convinced we absolutely require it...

Patch-set: 7