Update patch set 7
Patch Set 7:

> do we automagically diff in some place?

I don't think so, though quick-diff (not in this change) can be rather auto-magic. But the point is that LFS pointers are not discriminated in the UI somehow (and cannot easily be made so), so the user unsuspectingly triggers a comparison, and boom, it'll start downloading a whole DVD image or whatever.

Patch-set: 7