Update patch set 7
Patch Set 7:

> > do we automagically diff in some place?
 > 
 > I don't think so, though quick-diff (not in this change) can be
 > rather auto-magic. But the point is that LFS pointers are not
 > discriminated in the UI somehow (and cannot easily be made so), so
 > the user unsuspectingly triggers a comparison, and boom, it'll
 > start downloading a whole DVD image or whatever.

If you tell Eclipse to diff a .iso file, and you don't have all LFS blobs locally already, yes. That's why I was thinking about the "emergency break". Let's ask differently: what would happen if I try to quick-diff a DVD image which is NOT in LFS, but in normal GIT? The "only" difference is the download, right? We can certainly handle this with some kind of timeout. Maybe a global option (quick-diff-lfs-download-timeout-whatever)? LFS would need to know the purpose of the download then though, which is not nice...

Patch-set: 7