Update patch set 7
Patch Set 7:

> > > do we automagically diff in some place?
 > >
 > > I don't think so, though quick-diff (not in this change) can be
 > > rather auto-magic. But the point is that LFS pointers are not
 > > discriminated in the UI somehow (and cannot easily be made so),
 > so
 > > the user unsuspectingly triggers a comparison, and boom, it'll
 > > start downloading a whole DVD image or whatever.
 > 
 > If you tell Eclipse to diff a .iso file, and you don't have all LFS
 > blobs locally already, yes. That's why I was thinking about the
 > "emergency break". Let's ask differently: what would happen if I
 > try to quick-diff a DVD image which is NOT in LFS, but in normal
 > GIT? The "only" difference is the download, right? We can certainly
 > handle this with some kind of timeout. Maybe a global option
 > (quick-diff-lfs-download-timeout-whatever)? LFS would need to know
 > the purpose of the download then though, which is not nice...

I shouldn't have said "DVD image", that was just a tongue-in-cheek placeholder for "very large file". (BTW, a 4GB download would take about 30-45min with my connection). Maybe we could build something with progress monitors and give the user a way to cancel... though I'm not sure; IStorage.getContents() doesn't take a progress monitor. Would have to look closely where such cancellation could be done in Eclipse's compare framework. But in general it's not nice if the user only notices after the fact that the operation will take minutes or longer, and then has to cancel. If we could know up front that an operation on some file would trigger expensive side-effects we could ask before, but in general we can't know that, and even special-casing this for LFS would be non-trivial and potentially slow down things in general. For this diff (comparison/blame/quick diff) business, a git attribute might be a flexible and relatively cheap way to give the user some control.

The text/binary attribute cannot be used; it may in fact be OK to do comparisons of binary files. Don't know if Eclipse has an image comparator, but it does at least try to determine itself whether some content is text or binary. It could check the first bytes, and if they both indicate images, could open an image diff viewer. (Image diffs can be done quite nicely with a variety of techniques (side-by-side view, or subtracting, or a blink comparator. Perhaps merging would be restricted to "take ours" or "take theirs"...) So for a 7MB jpg stored in LFS, it might be OK to smudge filter to do the download, for a 700MB mp4 video the user might not want it. So set +rawdiff on the mp4 video and it just won't happen. Of course the comparison then compares with the LFS pointer.

BTW, it's not just about download time. Even if the LFS blob is already cached locally, I fear that doing the smudge and comparing two huge files may overwhelm Eclipse (memory requirements; text diff on big files already is slow, warns about falling back to a less precise algorithm, and even then takes time).

Patch-set: 7
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0