Update patch set 7
Patch Set 7:

> I shouldn't have said "DVD image", that was just a tongue-in-cheek
 > placeholder for "very large file". (BTW, a 4GB download would take
 > about 30-45min with my connection). Maybe we could build something
 > with progress monitors and give the user a way to cancel... though
 > I'm not sure; IStorage.getContents() doesn't take a progress
 > monitor. Would have to look closely where such cancellation could
 > be done in Eclipse's compare framework. But in general it's not
 > nice if the user only notices after the fact that the operation
 > will take minutes or longer, and then has to cancel. If we could

A cheap (but hacky?) way would be wrapping filtering in a job which we join for while spinning the event loop.

 > know up front that an operation on some file would trigger
 > expensive side-effects we could ask before, but in general we can't
 > know that, and even special-casing this for LFS would be
 > non-trivial and potentially slow down things in general. For this
 > diff (comparison/blame/quick diff) business, a git attribute might
 > be a flexible and relatively cheap way to give the user some
 > control.

Changing built-in filters to be able to ask them whether a certain operation is expensive might be possible, but once we're down to native filter, we lost anyway... Yes, I think if we need/want that control, an attribute is the cheapest/best way to achieve it.

 > ...
 > BTW, it's not just about download time. Even if the LFS blob is
 > already cached locally, I fear that doing the smudge and comparing
 > two huge files may overwhelm Eclipse (memory requirements; text
 > diff on big files already is slow, warns about falling back to a
 > less precise algorithm, and even then takes time).

Yes, but in the case there is no download we're neither better nor worse than ANY comparison in Eclipse, right?

Patch-set: 7
Reviewer: Gerrit User 326 <326@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0