Update patch set 4
Patch Set 4: Code-Review-1

Is an alternative mechanism for the post-commit hook really necessary? 

If it is: provide icons as SVG (in the top-level icons directory), and as PNG in 16x16 and 32x32 resolutions in the plugin, with the 32pixel versions named *@2x.png. Also, is a MessageDialog with custom buttons not sufficient? IIRC, that one already provides error/warning/info icons.

Patch-set: 4
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1