[WIP] Adds repository selection combo to the fetch gerrit change dialog
Removes the message that no repository is selected if the fetch gerrit
change dialog is called without a repository selected (e.g. via ctrl +
3).
Instead there is a combo selector on top of the dialog that lets the
user choose what repository he wants (even if a repository is selected).

Bug: 507733
Change-Id: I0423a10270e9201cc8ebd1a077cf53c1e3aa288b
Signed-off-by: Jonas Hungershausen <jonas.hungershausen@vogella.com>