Speed up getting the last commit that changed a file
"Compare with HEAD" uses CompareUtils.getHeadTypedElement() to
get the commit and blob. However, this method not always uses the
HEAD revision: it actually looks up the last commit that changed
the file and uses that commit. This is done because it enables
us to display better commit and author info: we'll show the
author who indeed last touched a file, not the author of the
HEAD commit.

However, a RevWalk with a path filter by default tries to present
a cleaned history, rewriting commit parents to skip commits that
didn't change the files selected by the filter. That is an expensive
operation since it fully reads and caches the whole repository
history!

This parent rewriting is not needed at all when we're only interested
in the last commit that changed a file. Therefore, configure the
RevWalk not to rewrite parents.

On the large repository mentioned in the bug report, this brings down
the time to open a compare editor for a staged file from > 12 sec to
about 1 sec.

Bug: 531948
Change-Id: I47da3618be8d7aac9fb85ab45cfdbefa184a3c58
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>