Update patch set 1
Patch Set 1:

Actually not so sure anymore. I found the "why" the version was not in our builds. The reason was that we have our own gerrit repository we're building from, and we did not push the release tags there.

Gerrit build uses 'git describe --always --match "v[0-9].*" --dirty' to calculate the release number.

Patch-set: 1