Show affected projects when deleting repository
Have a list of projects being deleted underneath the checkbox for
confirming project deletion. The height is fixed to 100 pixels (6 lines
on my machine), vertical scrolling is possible after checking the
checkbox. A screenshot is in bugzilla.

This changes the API of the RemoveCommand, I'm not sure what that means
for versioning.

Change-Id: I7f04c947919fa56cbc0f44a108a336c881b9858d
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Bug:478159