Update patch set 1
Patch Set 1:

(4 comments)

I implemented the "match at the front if contains wildcard" behavior. During tests that feels natural. From a viewpoint of symmetry we could also anchor at the end if a wildcard is there, but to me that seems counterintuitive for normal usage, since it would require an input of "*partialName*" instead of "*partialName" to match at all.

Patch-set: 1
Label: Verified=0