Improve check for uncommitted changes before cherry-pick
First, give the dialog an appropriate title. The default title
mentions rebasing, which is confusing when the user tries to
cherry-pick.

Second, limit the check to files in the diff to be applied. This
allows the cherry-pick to proceed successfully in a lot more cases.
The behavior is then more in line with rebasing or checking out,
which also can succeed if the working tree is dirty.

Bug: 470452
Change-Id: I0ea3985ecebb8383f45f560e94f3cde1309d2570
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>