Update patch set 3
Patch Set 3:

Just remembered that we also have bug 521461 -- for which we may need a way to detect the Gerrit version. So probably removing the check altogether is counter-productive: we'll have to re-add something similar later. Unless we could do for bug 521461 a more targeted feature test instead of a version test...

Sorry I didn't think of that sooner.

Patch-set: 3