Update patch set 1
Patch Set 1:

(3 comments)

Not sure about the claim this had recently become popular. Googling for the term gives a measly ~3000 hits, not all of which are related to this use. But allright, _some_ people might want to use this.

A pity that the commit message editor doesn't visualize the emojis... There was an attempt to make the commit editor component pluggable in https://git.eclipse.org/r/#/c/65199/ . But that has its own set of design issues... But if those were solved, a markdown editor could do so.

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0