NewRepositoryWizard: suggest complete repository path
This simplifies the new repository wizard by suggesting "repository",
"repository2", ... as the default name of the new repository. That way
the wizard is even more simple to use for end users.

This also simplifies validation, since with the old behaviour this could
easily trigger warnings of the directory not being empty (which is now
guaranteed).

Bug: 534404
Change-Id: I3175f8f0e791efa63ba458a250c5fa81c5164504
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>