Update patch set 4
Patch Set 4:

This implements the missing RepositoryChangeScanner (the one that periodically checks for git changes done outside of EGit/JGit). It also skips ignored directories.

During my tests on a Windows machine (which showed that this indeed improves performance quite a bit) I also noticed a problem with the content provider for the Git Repositories view being slow. I may revisit that in a later commit.

Patch-set: 4
Label: Verified=0