Natural sort for tags in CreateTagDialog
Give the tag viewer a comparator, and change the label provider
from an ITableLabelProvider to a simple LabelProvider. The viewer
has one column only anyway, and with the previous label provider
implementation, the comparer wouldn't work. The viewer would use
getColumnText(), but the ViewerComparator would use the inherited
getText() from WorkbenchLabelProvider, which always returns the
empty string since the elements don't adapt to IWorkbenchAdapter.

Bug: 529449
Change-Id: Ib1371f13c80334a9f41246fc96134f3c99198fb4
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>