Do not display "[NO-HEAD]" for empty repository
After a new repository has been created, its "master" branch is an
unborn branch. EGit showed "NO-HEAD" in the decorations for this case.

This is confusing for novices and looks like something went wrong. The
canonical implementation shows after "git init; git status": "On branch
master" and "Initial commit".

So let's display "master" in this case, too. Reserve the "NO-HEAD" for
cases where there is really no HEAD.

Also show such a HEAD in the repositories view under "References" and
decorate it as "[master] unborn; will be created by commit". Limit
showing such a HEAD to the repositories view: the content provider is
used in other places where such a HEAD node would only cause confusion
or malfunctioning. Use an extension factory to create the content
provider for the repositories view; in this way the new behavior can be
enabled there only without having to touch all the other places where
this content provider is used. Adapt the label provider to handle the
new case, and ensure that the history page doesn't get confused and
broken if HEAD is such an unborn branch.

Bug: 533596
Change-Id: Ifd0547a156f0a5616ce7e032ea8daebd2596f31e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>