Update patch set 1
Patch Set 1:

@Michael: sorry, don't have the time. I start my runtime workbench; select one of my test repos. Staging view fills, including a previously stored commit message. I open Preferences->Team->Git->Configuration, change the user.name, click "Apply", and then several things happen:

* The committer is changed
* The commit message is cleared
* The author is changed.

Anyway, I did spend some time over the week-end to make that listener do just what it should do: change the committer, and the author, if author.equals(committer).

In doing so I found a couple of bugs in the CommitMessageComponent.

I do have something working now, but this will need UI tests. Those will take me quite a bit longer to write.

So I suggest we do revert this for now on 5.0 (and, via merging, also on master), and then re-do this properly later again.

Patch-set: 1