Staging view: update author/committer on config changes
When the config changes, update the committer (and possibly also a
signed-off line in the commit message) to the possibly new user.name
and user.email values.

If the author is the same as the committer, also update the author of
the commit.

Correct the logic in CommitMessageComponent for updating the sign-off
settings. Previously, editing the committer field would switch off
the "sign off" flag if the committer was invalid (quite common if you
type a new value), and thus the user had to re-enable it manually and
correct an old or wrong signed-off-by line explicitly.

Bug: 533019
Change-Id: I13252da7beec6cd1ecd1ad2831b4d85c94d46636
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>