Update patch set 1
Patch Set 1:

> I started from a clean ~/.gitconfig having no LFS related
 > configuration.

Should be perfectly fine, yes.

 > 
 > * The button to enable LFS under "Preferences > Team > Git >Enable
 > LFS support globally (for the current user)"
 > is disabled and the title of this option says "LFS support (not
 > installed)" though I have installed the org.eclipse.jgit.lfs
 > bundle.

I can reproduce although I don't understand. Likely something has changed in the meantime which did not cause a conflict but malfunction.

 > * I cannot find any command in the repository context menu (in
 > repositories view) which allows to enable LFS for a repository.

It seems that the LFS bundle is somehow not detected, so this menu is missing as well.

 > * Switching on the option "Preferences > Team > Git > Confirmations
 > > Automatically configure LFS without confirmation" also doesn't
 > seem to make any difference.

I think this is all related to the same issue.

 > 
 > Could you add some documentation how to use this built-in LFS
 > integration to the user guide [1] ?
 > 
 > [1] https://wiki.eclipse.org/EGit/User_Guide

Once it's working and we merged (or are planning to merge) the change - yes, of course.

Patch-set: 1
Label: Verified=0