Update patch set 2
Patch Set 2:

owh well. seems that it does not yet play that nice with github. the auto detection for the LFS URI is implemented for Gerrit mostly (SSH based auto-login to the LFS service).

This is a JGit issue, not EGit. I prepared https://git.eclipse.org/r/124355 - it contains some first fixes which should make it work via SSH but not behind a HTTP proxy - this definitely needs more love/work (but again, it's not the EGit integration :)).

Patch-set: 2
Label: Verified=0