Update patch set 5
Patch Set 5:

@Markus: 
can you please document how to use this in the EGit User Guide ?
Ensure to mention that it works with ssh but there is still some work to do for http/https.

Let me know if you can finish that today then we can include this documentation in 5.0.

Patch-set: 5