Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure if there is some kind of policy about this in egit, but every other piece of code in the 3 git related repositories uses jgit.annotations, not jdt.annotations.

Independent of that, if you want to use jdt.annotations, those should not be in the manifest at all, but instead in build.properties as additional bundle. That way they are available at compile time, but do not become part of the plugin interface.

In fact, egit.mylyn.ui already has that entry, so the manifest entry can be deleted immediately. The only thing that is left open is cleaning up the target definition afterwards. I see both jdt.annotations 1.x and 2.x in the target, and the first should be removed.

Patch-set: 1
Reviewer: Gerrit User 440 <440@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1