Update patch set 1
Patch Set 1:

> The temporary update site still contains the wrong HttpClient
 > version.
 
the httpclient is taken from the jgit p2 repository, since the corresponding change for jgit is still in review the egit build uses the p2 repository from the last build on the stable-5.0 branch which still brings the new httpclient version. Will retrigger this build.

 > And the build log shows that for some reason, only the Photon Orbit
 > site was used: http://download.eclipse.org/tools/orbit/downloads/drops/R20180606145124/repository

this shouldn't make a difference since the httpclient is carried into the p2 repository from the jgit p2 repository

Patch-set: 1
Label: Verified=0