History search: use CSS for text field background color
Use a CSS class on the text field to set the light red background
color when there are no matches. Remove the CSS class to reset the
background. This makes the background color changes work as expected
with themes that define their own background.

Note that changing the background of a text field is not supported
on all platforms by SWT. (Currently only gtk and Windows.)

Bug: 532940
Change-Id: Ieb1c6b70a5df9b23c18b501b89379fa9cdeb66fb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>