Update patch set 3
Patch Set 3: Code-Review+1

Not selecting is fine for me.

I wish reveal() were a bit cleverer such that it'd show the parents if it could while still showing the actual node being revealed.

(What I mean is: make the dialog small enough to not show all entries, but still having more than two lines. Scroll to the bottom. Add entry "aaa.bbb=ccc". The TreeViewer shows as the top line "bbb=ccc" on my Mac. Would be nicer if it showed "aaa" as the top line and "bbb=ccc" as the second line.)

But that should be improved in TreeViewer.reveal(), not here.

Patch-set: 3