Update patch set 1
Patch Set 1: Code-Review+1

This solves the "two commits created on double click". I was still able to get on a double click a commit followed by a "can't commit; no staged changes" dialog.

I suspect that resolving this properly would require making sure that the buttons are enabled only when committing is at all possible (accounting for the case for amending, where it's OK to commit when there are no staged changes). And that whole logic might need revisiting if and when we get https://git.eclipse.org/r/#/c/77213/ fixed.

Patch-set: 1
Label: Code-Review=+1