Update patch set 1
Patch Set 1:

I'm reluctant to merge this.

On Mac, with a newly installed Eclipse Photon provisioned via Oomph, I get when launching a runtime Eclipse with the egit-4.7 target platform:

* tons of warnings about duplicate bundles (with identical version numbers)
* tons of warnings about bundle start failures (another singleton already active); may be follow-up warnings.

This doesn't happen with the old target platform.

The runtime Eclipse instance comes up fine in the end, but something is definitely strange.

Patch-set: 1