Update patch set 1
Patch Set 1:

With the egit-4.8 target, this doesn't happen, but there are other problems.

It appears that the mylyn version 3.24 that gets installed requires an Apache Lucene [6.0, 6.3.0), but there is only a Lucene 7.1. Moreover, there's an org.apache.lucene.queryparser 6 instead of 7 in the target platform.

No idea what's going on.

Patch-set: 1