Update patch set 1
Patch Set 1:

On Photon (on Mac), a started egit-4.7 runtime workbench sometimes even won't stop when it is quit normally. When that happens, Apache felix loops endlessly getting interrupted in its termination loop and fills the log with stack traces about it.

Maybe something is broken on my machine...

Patch-set: 1