Update patch set 1
Patch Set 1:

How does that change the read-only behaviour of file revisions for the user experience? If a temp file is put on disk, we may get an external editor which is not read-only, while the Eclipse internal editor is always read-only for revisions, or isn't it?

At least for me having read-only editors has been helpful several times when I accidentially tried editing in an old revision...

Patch-set: 1
Label: Code-Review=0