Update patch set 1
Patch Set 1: Code-Review+1

Using Eclipse Oxygen 3a as development Eclipse, starting an egit-4.7 runtime Eclipse works normally for me. So it really appears that the duplicate warnings and the fact that Apache Felix sometimes hangs are problems related to using a Photon workbench for development.

So I guess this would be OK to merge. (Provided the TPDs didn't change in the meantime, but I don't think they have.)

Patch-set: 1
Label: Code-Review=+1