Update patch set 3
Patch Set 3:

> Yes. Thomas, could you continue on this?

I'll look at it when I have the time. I have no repository with which I could reproduce the 8-second block you experience, so I will have to rely on you for testing. Unless you could make that 15G repo available to me... but as it's not public, that would probably need a full-blown contract between the owner and my employer for fixing this, with non-disclosure agreements, pricing, and everything. Much easier if you were available to test my changes against that repo.

But going back to your stack trace at https://bugs.eclipse.org/bugs/show_bug.cgi?id=485743#c17 I do not understand how you ended up concluding that this showHead/showTag/etc was the culprit. The block shown in the stack trace is _inside_ the asyncExec() within showCommitList(). Probably that asyncExec() executes concurrently with the next run of the GenerateHistoryJob, and gets blocked there. That wouldn't be fixed by this change anyway.

Patch-set: 3