Avoid that loading the history blocks updating the commit table
Somehow setting the input on the commit graph table got blocked on
the monitor of the SWTCommitList. The only other place where that
monitor can be held is in the GenerateHistoryJob itself.

However, CommitGraphTable.setInput() is always called not only with
that SWTCommitList but also its contents copied into an array. Thus
we can use this array copy instead of the list itself to get the
contents. On the array, no synchronization is needed.

That also means that inside the GenerateHistoryJob no synchronization
on that list is needed anymore, since only that job accesses it.

Bug: 485743
Change-Id: I77007b159d8b3de0a346b7cea69599fd61d80585
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>