Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> I'm wondering if also other people would like to have a 4th option in the merge tool selection: Always ask. Basically I'm completely happy with almost always using the Eclipse internal diff, but just sometimes I would like to use Beyond Compare instead.
> 
> However, I'm not sure if it is worth the effort to investigate this _now_. It is probably better to just get the current implementation stable and test drive it for a while, to see if I would still have that need or not.

I think you are talking about the EGit usability, so to have an option "Always ask" before starting internal or external diff/mergetool? I think it would be good. For you information this change will not be merged soon, because the whole logic without the UI code will be implemented / moved to JGit and then i will reuse this or create new change for EGit UI integration based at JGit API. I've just rebased it to check if it still works with EGit 5.x... i hope to continue with this after git worktree is finished in next weeks...

Patch-set: 13
Label: Verified=0