Update patch set 1
Patch Set 1:

(2 comments)

Looks nice, but the text fields are now slightly higher and the text is no longer vertically centered. That's quite visible in the screenshot you attached to bug 537200 (Windows?); it's less pronounced but still a noticeable difference on Mac OS X.

See inline comments for how to fix. (Default vertical alignment is SWT.FILL, but the grid cell is now a bit higher because of the icon or the new container around icon and text label.)

Patch-set: 1
Label: Code-Review=0