Update patch set 2
Patch Set 2:

Thomas, I've applied your changes and removed the additional grouping composite to avoid that its spacing and margin get relevant.

Additionally I've removed the margin that was previously around the commit buttons and the committer/author fields. Previously the "A" of "Author" and the end of the author input field where not in line with the left and right border of the commit message field, but now they are (at least on my Windows system).

Patch-set: 2
Label: Verified=0