History table: adapt column widths when view size changes
Use a TableColumnLayout; give columns minimum widths, and give the
commit ID column a fixed size. Re-layout the table when the width
changes, and if the table then doesn't fit the view, re-distribute
all column widths afresh.

This results in the table always using at least the available width,
and getting a horizontal scroll bar only if the user makes columns so
wide that they cannot fit the view. (Or makes the view that small.)

Bug: 537181
Change-Id: Ic105acd0ffc7877e8834888b841e3f0aa96993cb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>