Add support for multi repository selection for switching branches
When multiple repositories are selected the Switch To... option
will display only the branches that are present in all of the
repositories. By clicking in any of them, the user will concurrently
switch branches for all the repositories. If the list of branches is
to be empty, a disabled menu item will be added to show that no
matching branches were found.

Change-Id: I0f850baeb94cb8cb4169d8684bb37756b5c0c372
Signed-off-by: Luís Copetti <lhcopetti@gmail.com>