Update patch set 3
Patch Set 3:

(2 comments)

> I was quite happy
 > before about the 2 different decorations showing the branch name
 > and working dir in different colors. I'm wondering if this can
 > somehow be reproduced

Done. I managed to reproduce this, but it's a bit hacky. Essentially it means taking this bit out of the new lightweight decorator and moving it back into the label provider. Lightweight decorators have no control over the styling of their decorations. (Which is a hole in the whole concept of lightweight decorators.)

I discovered that I also had to add a cache to avoid flickering of decorations in the UI. For text decorations, I've found a way, but not for image decorations. If that bothers anyone, we'd have to move the image decoration part back into the label provider, but that means we'd also again execute potentially slow operations (resolving refs) in the UI thread.

Patch-set: 3
Label: Verified=0