Update patch set 6
Patch Set 6:

> https://ci.eclipse.org/egit/job/egit.gerrit/10676/ : UNSTABLE

So it's not a timeout problem. So far I cannot reproduce this locally on OS X. It's a bit strange: the repos appear to be gone, but somehow index diff cache entries for them are still around, and the ones reported are from tests that use the label provider in wizards. I already verified that dispose() is called on these label providers, and at least on OS X it is.

Patch-set: 6