Update patch set 1
Patch Set 1:

(1 comment)

It would be really cool if we had additionally a little context menu on these commits and file diffs with those operations available. I had never noticed that I could open a fetched commit by double-clicking...! Perhaps add that in a follow-up commit. Ctrl/Cmd-C (Copy) of selected items (copying the selected label texts to the clipboard) would also be nice to have.

In general the discoverability of such operations in dialogs is a bit of a problem. Even with a context menu a user has no indication that such a thing would be available. I don't have a bright idea how to improve that... The help on that dialog is also not exactly helpful; it goes to some very generic Eclipse help page. (But that's an entirely different topic. In general the help system of EGit is woefully underdeveloped.)

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0