Update patch set 3
Patch Set 3:

(2 comments)

Thanks for the quick review, I appreciated it. By the way, I also noticed that I had not considered what would happen when one or more of the user's multi-selection consisted of a non-repository resource. I fixed that by applying a stricter property filter so that the 'Switch To...' option will not be visible when that happens.

Patch-set: 3