Property sheet support for tags
When selecting a tag in the repository view, the properties view now
shows
* name and id of the commit the tag points to
* data about the tag itself

Properties are grouped by categories.

For annotated tags, the tagger is shown using sub-nodes for name,
e-mail, and the date formatted according to the EGit date preferences.

When a tag's properties are shown, the property page refreshes on date
format changes.

Deferred to follow-up changes:
* Showing information about the signature of signed tags; that needs
  signature verification in JGit first.
* Showing the full tag message if different from the short message.

Bug: 537695
Change-Id: Ie8c3a51a4f07795a2d28b561ce2d53668b20c723
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>