Property sheet support for tags
When selecting a tag in the repository view, the properties view now
shows information about the tag and its target (recursively in case
the tag points at another tag).

The object pointed to is shown as a nested property, which itself is
expandable if it's a tag or commit. PersonIdents for taggers and also
for commit authors and committers are show also as expandable nested
properties, with sub-nodes for name, e-mail, and the date formatted
according to the EGit date preferences.

When a tag's properties are shown, the property page refreshes on date
format changes.

Deferred to follow-up changes:
* Showing information about the signature of signed tags; that needs
  signature verification in JGit first.
* Showing the full tag or commit message if different from the short
  message.

Bug: 537695
Change-Id: Ie8c3a51a4f07795a2d28b561ce2d53668b20c723
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>